+++ R/pas_createNew.R
@@ -69,7 +69,7 @@
 #'
 #' pas <-
 #'   pas_createNew(
-#'     api_key = MY_API_READ_KEY,
+#'     api_key = PURPLE_AIR_API_READ_KEY,
 #'     countryCodes = "US",
 #'     stateCodes = "WA",
 #'     counties = c("Okanogan", "Ferry"),