+++ R/pat_createNew.R
@@ -38,7 +38,7 @@
 #'
 #' pat <-
 #'   pat_createNew(
-#'     api_key = MY_API_READ_KEY,
+#'     api_key = PURPLE_AIR_API_READ_KEY,
 #'     pas = MY_PAS,
 #'     sensor_index = "76545",
 #'     startdate = "2023-01-01",
@@ -241,7 +241,7 @@ pat_createNew <- function(
 
 if ( FALSE ) {
 
-  api_key = MY_API_READ_KEY
+  api_key = PURPLE_AIR_API_READ_KEY
   pas = example_pas
   sensor_index = "76545"
   startdate = "2023-01-01"