+++ local_jon/pa_API_interactive_test.R
@@ -5,7 +5,7 @@ source("global_vars.R")
 # ===== Keys ===================================================================
 
 pa_checkAPIKey(
-  api_key = MY_API_READ_KEY
+  api_key = PURPLE_AIR_API_READ_KEY
 )
 
 # ===== Sensors ================================================================
@@ -19,12 +19,12 @@ end <-
   as.numeric()
 
 a <- pa_getSensorData(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   sensor_index = 896
 )
 
 a <- pa_getSensorHistoryCSV(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   sensor_index = 896,
   start_timestamp = start,
   end_timestamp = end,
@@ -32,7 +32,7 @@ a <- pa_getSensorHistoryCSV(
 )
 
 a <- pa_getSensorHistory(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   sensor_index = 896,
   start_timestamp = start,
   end_timestamp = end,
@@ -41,7 +41,7 @@ a <- pa_getSensorHistory(
 )
 
 a <- pa_getSensorsData(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   location_type = 0,
   max_age = 3600 * 3,
   nwlng = -125,
@@ -53,50 +53,50 @@ a <- pa_getSensorsData(
 # ===== Groups =================================================================
 
 pa_createGroup(
-  api_key = MY_API_WRITE_KEY,
+  api_key = PURPLE_AIR_API_WRITE_KEY,
   name = "Ballard"
 )
 
 pa_createMember(
-  api_key = MY_API_WRITE_KEY,
+  api_key = PURPLE_AIR_API_WRITE_KEY,
   sensor_index = NULL
 )
 
 pa_deleteGroup(
-  api_key = MY_API_WRITE_KEY,
+  api_key = PURPLE_AIR_API_WRITE_KEY,
   group_id = 1652
 )
 
 pa_deleteMember(
-  api_key = MY_API_WRITE_KEY,
+  api_key = PURPLE_AIR_API_WRITE_KEY,
   group_id = 505,
   member_id = 3855
 )
 
 pa_getGroupsList(
-  api_key = MY_API_READ_KEY
+  api_key = PURPLE_AIR_API_READ_KEY
 )
 
 pa_getGroupDetail(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   group_id = 505
 )
 
 a <- pa_getMemberData(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   group_id = 505,
   member_id = 3856
 )
 
 a <- pa_getMemberHistory(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   group_id = 505,
   member_id = 3856,
   average = 0
 )
 
 a <- pa_getMemberHistory(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   group_id = 505,
   member_id = 3856,
   start_timestamp = start,
@@ -105,7 +105,7 @@ a <- pa_getMemberHistory(
 )
 
 a <- pa_getMembersData(
-  api_key = MY_API_READ_KEY,
+  api_key = PURPLE_AIR_API_READ_KEY,
   group_id = 505
 )